make catalog/helm and catalog/git utility packages
