+++ pkg/catalog/manager/catalog_sync.go
@@ -1,7 +1,7 @@
 package manager
 
 import (
-	helmlib "github.com/rancher/rancher/pkg/catalog/helm"
+	helmlib "github.com/rancher/rancher/pkg/helm"
 	v3 "github.com/rancher/rancher/pkg/generated/norman/management.cattle.io/v3"
 	"github.com/rancher/rancher/pkg/namespace"
 	"github.com/sirupsen/logrus"
