+++ pkg/controllers/management/kontainerdrivermetadata/url.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/rancher/norman/types/convert"
-	"github.com/rancher/rancher/pkg/catalog/git"
+	"github.com/rancher/rancher/pkg/git"
 	"github.com/rancher/rancher/pkg/settings"
 	"github.com/rancher/rke/types/kdm"
 	"github.com/rancher/wrangler/pkg/randomtoken"
