+++ pkg/git/git.go
@@ -5,14 +5,14 @@ import (
 	"fmt"
 	"net/url"
 	"os/exec"
+	"regexp"
 	"strings"
 
 	"github.com/pkg/errors"
-	catUtil "github.com/rancher/rancher/pkg/catalog/utils"
 )
 
 func Clone(path, url, branch string) error {
-	if err := catUtil.ValidateURL(url); err != nil {
+	if err := ValidateURL(url); err != nil {
 		return err
 	}
 	return runcmd("git", "clone", "-b", branch, "--single-branch", url, path)
@@ -32,7 +32,7 @@ func HeadCommit(path string) (string, error) {
 }
 
 func RemoteBranchHeadCommit(url, branch string) (string, error) {
-	if err := catUtil.ValidateURL(url); err != nil {
+	if err := ValidateURL(url); err != nil {
 		return "", err
 	}
 	cmd := exec.Command("git", "ls-remote", url, branch)
@@ -48,7 +48,7 @@ func RemoteBranchHeadCommit(url, branch string) (string, error) {
 }
 
 func IsValid(url string) bool {
-	if err := catUtil.ValidateURL(url); err != nil {
+	if err := ValidateURL(url); err != nil {
 		return false
 	}
 	err := runcmd("git", "ls-remote", url)
@@ -77,8 +77,21 @@ func FormatURL(pathURL, username, password string) string {
 }
 
 func CloneWithDepth(path, url, branch string, depth int) error {
-	if err := catUtil.ValidateURL(url); err != nil {
+	if err := ValidateURL(url); err != nil {
 		return err
 	}
 	return runcmd("git", "clone", "-b", branch, "--single-branch", fmt.Sprintf("--depth=%v", depth), url, path)
+}
+
+var (
+	controlChars   = regexp.MustCompile("[[:cntrl:]]")
+	controlEncoded = regexp.MustCompile("%[0-1][0-9,a-f,A-F]")
+)
+
+func ValidateURL(pathURL string) error {
+	// Don't allow a URL containing control characters, standard or url-encoded
+	if controlChars.FindStringIndex(pathURL) != nil || controlEncoded.FindStringIndex(pathURL) != nil {
+		return errors.New("Invalid characters in url")
+	}
+	return nil
 }
